Fix build failure on Win64 since r14458.
R=jkummerow@chromium.org

Review URL: https://codereview.chromium.org/14143008

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@14463 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/frames-inl.h
@@ -236,7 +236,7 @@ inline int JavaScriptFrame::ComputeOperandsCount() const {
   ASSERT(IsAligned(stack_size_in_bytes, kPointerSize));
   ASSERT(type() == JAVA_SCRIPT);
   ASSERT(stack_size_in_bytes >= 0);
-  return stack_size_in_bytes >> kPointerSizeLog2;
+  return static_cast<int>(stack_size_in_bytes >> kPointerSizeLog2);
 }
